+++ .gitignore
@@ -1,4 +1,6 @@
 .DS_Store
+sent_emails/
+media/
 # Byte-compiled / optimized / DLL files
 __pycache__/
