+++ main/templates/main/note_confirm_delete.html
@@ -9,7 +9,7 @@
     <p class="text-center">Are you sure you want to delete this note ? " {{ note }} "</p>
     <div class="flex gap-2 py-4 justify-center">
     <input type="submit" value="Yes" class="py-2 px-4 bg-red-500 text-white">
-    <a href="{% url 'notes' %}" class="py-2 px-4 bg-gray-300 text-white">No</a>
+    <a href="{% url 'main:notes' %}" class="py-2 px-4 bg-gray-300 text-white">No</a>
     </div>
   </form>
