+++ main/templates/main/note_detail.html
@@ -5,13 +5,13 @@
 
 <div class="p-4 space-y-4">
   <h1 class="text-2xl font-semibold">
-    {{ note.title|title }} <span class="text-base font-thin"> by <a href="{% url 'profile' note.user.id %}">{{ note.user|title }}</a> | {{ note.created_time|date:'N j, Y' }}</span>
+    {{ note.title|title }} <span class="text-base font-thin"> by <a href="{% url 'users:profile' note.user.id %}">{{ note.user|title }}</a> | {{ note.created_time|date:'N j, Y' }}</span>
   </h1>
   <p>{{ note.note|safe }}</p>
 
   {% if note.user == user %}
   <div class="my-4">
-    <a href="{% url 'note-update' note.id %}" class="block w-fit mx-auto px-4 py-2 bg-gray-400 hover:bg-gray-600 rounded-xl text-white">Edit</a>
+    <a href="{% url 'main:note-update' note.id %}" class="block w-fit mx-auto px-4 py-2 bg-gray-400 hover:bg-gray-600 rounded-xl text-white">Edit</a>
   </div>
   {% endif %}
 </div>
