+++ main/templates/main/note_form.html
@@ -32,9 +32,9 @@
 
       <div class="py-4 mt-6 mb-10 flex gap-4 justify-center">
         <input type="submit" value="Save" class="px-4 py-2 bg-gradient-to-tr hover:bg-gradient-to-bl from-cyan-500 to-blue-500 text-white rounded-xl">
-        <a href="{% url 'notes' %}" class="bg-gray-400 hover:bg-gray-600 text-white rounded-xl px-4 py-2">Cancel</a>
+        <a href="{% url 'main:notes' %}" class="bg-gray-400 hover:bg-gray-600 text-white rounded-xl px-4 py-2">Cancel</a>
         {% if page_title != "Create" %}
-          <a href="{% url 'note-delete' note.id %}" class="bg-red-500 hover:bg-red-700 text-white rounded-xl px-4 py-2">
+          <a href="{% url 'main:note-delete' note.id %}" class="bg-red-500 hover:bg-red-700 text-white rounded-xl px-4 py-2">
             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="24" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
               <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5Zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6Z"/>
               <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1ZM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118ZM2.5 3h11V2h-11v1Z"/>
