+++ main/views.py
@@ -1,45 +1,16 @@
-from django.shortcuts import render, redirect
-from django.views.generic import ListView, CreateView, UpdateView, DeleteView, FormView, TemplateView
+from django.views.generic import ListView, CreateView, UpdateView, DeleteView, TemplateView
 from django.urls import reverse_lazy
-
-from django.contrib.auth.views import LoginView
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth import login
+
 from django.http import HttpResponseForbidden
 
 from django.db.models import Q
 
 from .models import Note
 
-from .forms import NoteForm, RegisterForm
+from .forms import NoteForm
 
-class RegisterPage(FormView):
-    template_name = 'main/register.html'
-    form_class = RegisterForm
-    redirect_authenticated_user = True # we want to redirect authenticated users if trying to access the register page -> doesn't seem to work
-    success_url = reverse_lazy('notes')
 
-    def form_valid(self, form):
-        # we override the form_valid function so it logs the user in after registering
-        user = form.save()
-        if user is not None:
-            # if the user form is successfully saved, go ahead and login the user
-            login(self.request, user)
-        return super(RegisterPage, self).form_valid(form)
-
-    def get(self, *args, **kwargs):
-        if self.request.user.is_authenticated:
-            return redirect('notes') #forces to redirect to main page when user is authenticated
-        return super(RegisterPage, self).get(*args, **kwargs) #otherwise we apply the original method
-
-class CustomLoginView(LoginView):
-    template_name = 'main/login.html'
-    fields = '__all__'
-    redirect_authenticated_user = True
-
-    def get_success_url(self):
-        return reverse_lazy('all-notes') #force the login to redirect to notes page instead of profile (that doesn't exist)
 
 class NoteList(LoginRequiredMixin, ListView):
     model = Note
@@ -87,7 +58,7 @@ class NoteCreate(LoginRequiredMixin, CreateView):
     model = Note
     form_class = NoteForm
     #fields = ['title', 'note'] # we don't show 'user' -> the selection is overridden anyway (by the form valid function below)
-    success_url = reverse_lazy('notes')
+    success_url = reverse_lazy('main:notes')
 
     def form_valid(self, form):
     # forcing the form to have the authenticated user as author for the notes he creates
@@ -104,7 +75,7 @@ class NoteCreate(LoginRequiredMixin, CreateView):
 class NoteUpdate(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
     model = Note
     form_class = NoteForm
-    success_url = reverse_lazy('notes')
+    success_url = reverse_lazy('main:notes')
 
     def test_func(self):
         note_id = self.kwargs['pk']
@@ -123,4 +94,4 @@ class NoteUpdate(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
 class NoteDelete(LoginRequiredMixin, DeleteView):
     model = Note
     fields = '__all__'
-    success_url = reverse_lazy('notes')
+    success_url = reverse_lazy('main:notes')
