+++ users/templates/registration/password_reset_complete.html
@@ -1,0 +1,14 @@
+{% extends 'main/base.html' %}
+
+{% block title %}Reset Complete{% endblock %}
+
+{% block body %}
+
+  <div class="flex justify-center">
+    <h1 class="text-xl text-center font-semibold py-6">Your password has been set.  You may go ahead and log in now.</h1>
+  </div>
+  <div class="w-1/2 mx-auto">
+    <a href="{% url 'users:login' %}" class="flex w-fit mx-auto px-4 py-2 bg-gray-400 hover:bg-gray-600 text-white rounded-xl">Login</a>
+  </div>
+
+{% endblock body %}
