+++ users/templates/registration/password_reset_confirm.html
@@ -1,0 +1,41 @@
+{% extends "main/base.html" %}
+
+{% block title %}Confirm Reset{% endblock title %}
+
+
+{% block body %}
+
+{% if validlink %}
+<div class="flex items-center w-full">
+  <div class="w-80 p-8 m-4 mx-auto">
+    <h1 class="block w-full text-justify text-grey-darkest mb-6">Please enter your new password twice so we can verify you typed it in correctly.</h1>
+    <form method="post" class="mb-4">{% csrf_token %}
+      <input class="hidden" autocomplete="username" value="{{ form.user.get_username }}">
+      <div class="flex flex-col mb-4 py-2 px-3 ">
+        {{ form.new_password1.errors }}
+        <label for="id_new_password1">New password:</label>
+        {{ form.new_password1 }}
+      </div>
+      <div class="flex flex-col mb-4 py-2 px-3 ">
+        {{ form.new_password2.errors }}
+        <label for="id_new_password2">Confirm password:</label>
+        {{ form.new_password2 }}
+      </div>
+
+      <div>
+        <input type="submit" value="Change my password" class="block mx-auto px-4 py-2 bg-gray-400 hover:bg-gray-600 text-white rounded-xl">
+      </div>
+  </div>
+</div>
+
+
+</form>
+
+{% else %}
+
+<p>The password reset link was invalid, possibly because it has already been used.  Please request a new password reset.</p>
+
+{% endif %}
+
+
+{% endblock body %}
