+++ users/templates/registration/password_reset_done.html
@@ -1,0 +1,11 @@
+{% extends "main/base.html" %}
+
+{% block title %}Password Sent{% endblock title %}
+
+{% block body %}
+
+<p>We’ve emailed you instructions for setting your password, if an account exists with the email you entered. You should receive them shortly.</p>
+
+<p>If you don’t receive an email, please make sure you’ve entered the address you registered with, and check your spam folder.</p>
+
+{% endblock %}
