+++ users/templates/registration/password_reset_form.html
@@ -1,0 +1,26 @@
+{% extends "main/base.html" %}
+
+{% block title %}Password Reset{% endblock title %}
+
+{% block body %}
+<div class="p-4">
+  <h1 class="text-xl font-semibold py-6">Forgotten your password?</h1>
+  <p class="text-lg">Enter your email address below, and we’ll email instructions for setting a new one.</p>
+</div>
+
+<form method="post" class="dark:text-slate-600 space-y-4">{% csrf_token %}
+<fieldset class="module aligned">
+  <div class="grid gap-4 my-6">
+    <div class="space-x-4 mx-auto">
+        {{ form.email.errors }}
+        <label for="id_email" class="py-2">Email address:</label>
+        {{ form.email }}
+    </div>
+
+    <div class="my-6 mx-auto">
+      <input type="submit" value="Reset my password" class="px-4 py-2 bg-gray-400 hover:bg-gray-600 text-white rounded-xl">
+    </div>
+  </div>
+</fieldset>
+</form>
+{% endblock body %}
