+++ users/templates/users/register.html
@@ -27,7 +27,7 @@
 </form>
 
 <div class="py-4 text-center">
-  <p>Already have an acccount ? <a href="{% url 'login' %}">Login</a></p>
+  <p>Already have an acccount ? <a href="{% url 'users:login' %}">Login</a></p>
 </div>
