+++ users/templates/users/userprofile_form.html
@@ -65,7 +65,7 @@
 
     <div class="flex gap-4 w-1/2 mx-auto">
       <input type="submit" value="Save" class="px-4 py-2 bg-gradient-to-tr hover:bg-gradient-to-bl from-cyan-500 to-blue-500 text-white rounded-xl">
-      <a href="{% url 'profile' user.id %}" class="bg-gray-400 hover:bg-gray-600 text-white rounded-xl px-4 py-2">Cancel</a>
+      <a href="{% url 'users:profile' user.id %}" class="bg-gray-400 hover:bg-gray-600 text-white rounded-xl px-4 py-2">Cancel</a>
 
     </div>
   </form>
